Rename `core.py` to `model.py`
